Re-namespace machine managers (#471)
