+++ tests/Functional/Services/MachineManagerTest.php
@@ -2,14 +2,14 @@
 
 declare(strict_types=1);
 
-namespace App\Tests\Functional\Services\MachineManager;
+namespace App\Tests\Functional\Services;
 
 use App\Exception\MachineProvider\DigitalOcean\ApiLimitExceededException;
 use App\Exception\MachineProvider\DigitalOcean\HttpException;
 use App\Exception\MachineProvider\Exception;
+use App\Exception\MachineProvider\MachineNotFoundException;
 use App\Model\DigitalOcean\RemoteMachine;
-use App\Services\MachineManager\DigitalOceanMachineManager;
-use App\Services\MachineNameFactory;
+use App\Services\MachineManager;
 use App\Tests\AbstractBaseFunctionalTest;
 use App\Tests\Services\HttpResponseFactory;
 use DigitalOceanV2\Entity\Droplet as DropletEntity;
@@ -20,38 +20,37 @@ use GuzzleHttp\Handler\MockHandler;
 use GuzzleHttp\Psr7\Response;
 use Psr\Http\Message\ResponseInterface;
 use webignition\BasilWorkerManager\PersistenceBundle\Services\Factory\MachineFactory;
+use webignition\BasilWorkerManager\PersistenceBundle\Services\Factory\MachineProviderFactory;
 use webignition\BasilWorkerManagerInterfaces\Exception\MachineProvider\ExceptionInterface;
+use webignition\BasilWorkerManagerInterfaces\MachineActionInterface;
 use webignition\BasilWorkerManagerInterfaces\MachineInterface;
+use webignition\BasilWorkerManagerInterfaces\MachineProviderInterface;
+use webignition\BasilWorkerManagerInterfaces\ProviderInterface;
 use webignition\BasilWorkerManagerInterfaces\RemoteRequestActionInterface;
+use webignition\ObjectReflector\ObjectReflector;
 
-class DigitalOceanMachineManagerTest extends AbstractBaseFunctionalTest
+class MachineManagerTest extends AbstractBaseFunctionalTest
 {
     private const MACHINE_ID = 'machine id';
 
-    private DigitalOceanMachineManager $machineManager;
-    private MachineInterface $machine;
+    private MachineManager $machineManager;
     private MockHandler $mockHandler;
-    private string $machineName;
 
     protected function setUp(): void
     {
         parent::setUp();
 
-        $machineManager = self::$container->get(DigitalOceanMachineManager::class);
-        \assert($machineManager instanceof DigitalOceanMachineManager);
+        $machineManager = self::$container->get(MachineManager::class);
+        \assert($machineManager instanceof MachineManager);
         $this->machineManager = $machineManager;
 
         $machineFactory = self::$container->get(MachineFactory::class);
         \assert($machineFactory instanceof MachineFactory);
-        $this->machine = $machineFactory->create(self::MACHINE_ID);
-
-        $machineNameFactory = self::$container->get(MachineNameFactory::class);
-        \assert($machineNameFactory instanceof MachineNameFactory);
-        $this->machineName = $machineNameFactory->create(self::MACHINE_ID);
 
         $mockHandler = self::$container->get(MockHandler::class);
-        \assert($mockHandler instanceof MockHandler);
-        $this->mockHandler = $mockHandler;
+        if ($mockHandler instanceof MockHandler) {
+            $this->mockHandler = $mockHandler;
+        }
     }
 
     public function testCreateSuccess(): void
@@ -76,7 +75,7 @@ class DigitalOceanMachineManagerTest extends AbstractBaseFunctionalTest
         $expectedDropletEntity = new DropletEntity($dropletData);
         $this->mockHandler->append(HttpResponseFactory::fromDropletEntity($expectedDropletEntity));
 
-        $remoteMachine = $this->machineManager->create(self::MACHINE_ID, $this->machineName);
+        $remoteMachine = $this->machineManager->create($this->createMachineProvider());
 
         self::assertEquals(new RemoteMachine($expectedDropletEntity), $remoteMachine);
     }
@@ -95,9 +94,9 @@ class DigitalOceanMachineManagerTest extends AbstractBaseFunctionalTest
     ): void {
         $this->doActionThrowsExceptionTest(
             function () {
-                $this->machineManager->create(self::MACHINE_ID, $this->machineName);
+                $this->machineManager->create($this->createMachineProvider());
             },
-            RemoteRequestActionInterface::ACTION_CREATE,
+            MachineActionInterface::ACTION_CREATE,
             $apiResponse,
             $expectedExceptionClass,
             $expectedRemoveException
@@ -120,7 +119,7 @@ class DigitalOceanMachineManagerTest extends AbstractBaseFunctionalTest
         );
 
         try {
-            $this->machineManager->create(self::MACHINE_ID, $this->machineName);
+            $this->machineManager->create($this->createMachineProvider());
             self::fail(ExceptionInterface::class . ' not thrown');
         } catch (ExceptionInterface $exception) {
             self::assertEquals(
@@ -133,8 +132,6 @@ class DigitalOceanMachineManagerTest extends AbstractBaseFunctionalTest
     public function testGetSuccess(): void
     {
         $ipAddresses = ['10.0.0.1', '127.0.0.1', ];
-
-        self::assertSame([], $this->machine->getIpAddresses());
 
         $dropletData = [
             'networks' => (object) [
@@ -154,18 +151,23 @@ class DigitalOceanMachineManagerTest extends AbstractBaseFunctionalTest
         $expectedDropletEntity = new DropletEntity($dropletData);
         $this->mockHandler->append(HttpResponseFactory::fromDropletEntityCollection([$expectedDropletEntity]));
 
-        $remoteMachine = $this->machineManager->get(self::MACHINE_ID, $this->machineName);
+        $remoteMachine = $this->machineManager->get($this->createMachineProvider());
 
         self::assertEquals(new RemoteMachine($expectedDropletEntity), $remoteMachine);
     }
 
-    public function testGetMachineNotFound(): void
+    public function testGetThrowsMachineNotFoundException(): void
     {
         $this->mockHandler->append(HttpResponseFactory::fromDropletEntityCollection([]));
 
-        $remoteMachine = $this->machineManager->get(self::MACHINE_ID, $this->machineName);
+        $machineProvider = $this->createMachineProvider();
 
-        self::assertNull($remoteMachine);
+        self::expectExceptionObject(new MachineNotFoundException(
+            $machineProvider->getId(),
+            $machineProvider->getName()
+        ));
+
+        $this->machineManager->get($machineProvider);
     }
 
     /**
@@ -182,20 +184,20 @@ class DigitalOceanMachineManagerTest extends AbstractBaseFunctionalTest
     ): void {
         $this->doActionThrowsExceptionTest(
             function () {
-                $this->machineManager->get(self::MACHINE_ID, $this->machineName);
+                $this->machineManager->get($this->createMachineProvider());
             },
-            RemoteRequestActionInterface::ACTION_GET,
+            MachineActionInterface::ACTION_GET,
             $apiResponse,
             $expectedExceptionClass,
             $expectedRemoveException
         );
     }
 
-    public function testRemoveSuccess(): void
+    public function testDeleteSuccess(): void
     {
         $this->mockHandler->append(new Response(204));
-        $this->machineManager->remove(self::MACHINE_ID, $this->machineName);
 
+        $this->machineManager->delete($this->createMachineProvider());
         self::expectNotToPerformAssertions();
     }
 
@@ -204,16 +206,16 @@ class DigitalOceanMachineManagerTest extends AbstractBaseFunctionalTest
      *
      * @param class-string $expectedExceptionClass
      */
-    public function testRemoveThrowsException(
+    public function testDeleteThrowsException(
         ResponseInterface $apiResponse,
         string $expectedExceptionClass,
         \Exception $expectedRemoveException
     ): void {
         $this->doActionThrowsExceptionTest(
             function () {
-                $this->machineManager->remove(self::MACHINE_ID, $this->machineName);
+                $this->machineManager->delete($this->createMachineProvider());
             },
-            RemoteRequestActionInterface::ACTION_DELETE,
+            MachineActionInterface::ACTION_DELETE,
             $apiResponse,
             $expectedExceptionClass,
             $expectedRemoveException
@@ -225,10 +227,11 @@ class DigitalOceanMachineManagerTest extends AbstractBaseFunctionalTest
      */
     public function testExists(ResponseInterface $apiResponse, bool $expectedExists): void
     {
+//        ObjectReflector::setProperty($this->machine, Machine::class, 'remote_id', 123);
+
         $this->mockHandler->append($apiResponse);
 
-        $exists = $this->machineManager->exists(self::MACHINE_ID, $this->machineName);
-
+        $exists = $this->machineManager->exists($this->createMachineProvider());
         self::assertSame($expectedExists, $exists);
     }
 
@@ -265,9 +268,9 @@ class DigitalOceanMachineManagerTest extends AbstractBaseFunctionalTest
     ): void {
         $this->doActionThrowsExceptionTest(
             function () {
-                $this->machineManager->exists(self::MACHINE_ID, $this->machineName);
+                $this->machineManager->exists($this->createMachineProvider());
             },
-            RemoteRequestActionInterface::ACTION_GET,
+            MachineActionInterface::ACTION_EXISTS,
             $apiResponse,
             $expectedExceptionClass,
             $expectedRemoveException
@@ -275,7 +278,7 @@ class DigitalOceanMachineManagerTest extends AbstractBaseFunctionalTest
     }
 
     /**
-     * @param RemoteRequestActionInterface::ACTION_* $action
+     * @param MachineActionInterface::ACTION_* $action
      * @param class-string $expectedExceptionClass
      */
     private function doActionThrowsExceptionTest(
@@ -324,5 +327,63 @@ class DigitalOceanMachineManagerTest extends AbstractBaseFunctionalTest
                 'expectedRemoteException' => new ValidationFailedException('Bad Request', 400),
             ],
         ];
+    }
+
+    public function testFindRemoteMachineSuccess(): void
+    {
+        $dropletEntity = new DropletEntity([
+            'id' => 123,
+            'status' => RemoteMachine::STATE_NEW,
+        ]);
+
+        $this->mockHandler->append(HttpResponseFactory::fromDropletEntityCollection([$dropletEntity]));
+
+        $remoteMachine = $this->machineManager->findRemoteMachine(self::MACHINE_ID);
+
+        self::assertEquals(new RemoteMachine($dropletEntity), $remoteMachine);
+    }
+
+    /**
+     * @dataProvider findRemoteMachineThrowsMachineNotFoundExceptionDataProvider
+     *
+     * @param ResponseInterface[] $apiResponses
+     */
+    public function testFindRemoteMachineThrowsMachineNotFoundException(array $apiResponses): void
+    {
+        $this->mockHandler->append(...$apiResponses);
+
+        self::expectExceptionObject(new MachineNotFoundException(self::MACHINE_ID));
+
+        $this->machineManager->findRemoteMachine(self::MACHINE_ID);
+    }
+
+    /**
+     * @return array[]
+     */
+    public function findRemoteMachineThrowsMachineNotFoundExceptionDataProvider(): array
+    {
+        return [
+            'machine does not exist' => [
+                'apiResponses' => [
+                    HttpResponseFactory::fromDropletEntityCollection([]),
+                ],
+            ],
+            'request failed' => [
+                'apiResponses' => [
+                    new Response(503),
+                ],
+            ],
+        ];
+    }
+
+    private function createMachineProvider(): MachineProviderInterface
+    {
+        $machineProviderFactory = self::$container->get(MachineProviderFactory::class);
+        \assert($machineProviderFactory instanceof MachineProviderFactory);
+
+        return $machineProviderFactory->create(
+            self::MACHINE_ID,
+            ProviderInterface::NAME_DIGITALOCEAN
+        );
     }
 }
